[TAN-1716] Move to new ProjectIdeasVotesGenerator service
